+++ src/MicroOcpp/Core/Request.cpp
@@ -34,6 +34,10 @@ void Request::setOperation(std::unique_ptr<Operation> msg){
     operation = std::move(msg);
 }
 
+Operation *Request::getOperation(){
+    return operation.get();
+}
+
 void Request::setTimeout(unsigned long timeout) {
     this->timeout_period = timeout;
 }