+++ src/MicroOcpp/Model/ConnectorBase/ConnectorsCommon.cpp
@@ -10,6 +10,7 @@
 #include <MicroOcpp/Operations/ChangeAvailability.h>
 #include <MicroOcpp/Operations/ChangeConfiguration.h>
 #include <MicroOcpp/Operations/ClearCache.h>
+#include <MicroOcpp/Operations/DataTransfer.h>
 #include <MicroOcpp/Operations/GetConfiguration.h>
 #include <MicroOcpp/Operations/RemoteStartTransaction.h>
 #include <MicroOcpp/Operations/RemoteStopTransaction.h>
@@ -45,6 +46,8 @@ ConnectorsCommon::ConnectorsCommon(Context& context, unsigned int numConn, std::
         return new Ocpp16::ChangeConfiguration();});
     context.getOperationRegistry().registerOperation("ClearCache", [filesystem] () {
         return new Ocpp16::ClearCache(filesystem);});
+    context.getOperationRegistry().registerOperation("DataTransfer", [] () {
+        return new Ocpp16::DataTransfer();});
     context.getOperationRegistry().registerOperation("GetConfiguration", [] () {
         return new Ocpp16::GetConfiguration();});
     context.getOperationRegistry().registerOperation("RemoteStartTransaction", [&context] () {